+++ bitcoin/src/p2p/message_network.rs
@@ -11,7 +11,7 @@ use hashes::sha256d;
 use crate::consensus::{encode, Decodable, Encodable, ReadExt};
 use crate::internal_macros::impl_consensus_encoding;
 use crate::p2p::address::Address;
-use crate::p2p::constants::ServiceFlags;
+use crate::p2p::ServiceFlags;
 use crate::prelude::*;
 use crate::{io, p2p};
 
@@ -146,7 +146,7 @@ mod tests {
     use super::{Reject, RejectReason, VersionMessage};
     use crate::consensus::encode::{deserialize, serialize};
     use crate::internal_macros::hex;
-    use crate::p2p::constants::ServiceFlags;
+    use crate::p2p::ServiceFlags;
 
     #[test]
     fn version_message_test() {
